+++ src/phishbench/input/email_input/__init__.py
@@ -1,2 +1,7 @@
+"""
+
+"""
+from .email_input import read_dataset_email
+from .email_input import read_email_from_file
 from .models import EmailBody, EmailMessage, EmailHeader
 from .models.date_parse import parse_email_datetime